Add script to sync all the repos (if possible)
We'll need to improve and perhaps rethink this in the future. But
in the meantime running sync.sh in a regular fashion will push
the changes from the TransbankDevelopers/transbank-developers-docs
into the slate-tbk repository (with some automated transformations)
and then it will push those changes into the tbkdev_3.0-public
repository (using the build-all.sh script).

The sync will be aborted if any command fails. We should be able to
recover by re-running after solving any merge conflicts (if that's
the source of the error), or the script itself