Improve sync.sh to abort when new changes are pushed.
Since new changes mean that the CI platform will run a new build,
there is little point in going ahead.

I toyed with the idea of doing nothing if the last commit has
nothing to do with the upstream repo. But it's non-trivial logic
with bad consequences if I get it wrong. So we still run two full
builds for every upstream change (one because the upstream repo
changed and then one because the downstream repo was changed
by the sync itself). So there is room for future improvement :)